[artf791857][artf774132][COVERITY]: Fix code defects generated by coverity tool.
CID:61826.
Invalid Assignment: The type size of both side variables are different.
"fileSize" is 4 (unsigned int) and "fTellFileSize" is
8 (long)

+++ SN100x/halimpl/dnld/phDnldNfc.cc
@@ -1034,8 +1034,8 @@ NFCSTATUS phDnldNfc_LoadFW(const char* pathName, uint8_t** pImgInfo,
 *******************************************************************************/
 NFCSTATUS phDnldNfc_LoadBinFW(uint8_t** pImgInfo, uint32_t* pImgInfoLen) {
   FILE* pFile = NULL;
-  uint32_t fileSize = 0;
-  uint32_t bytesRead = 0;
+  long fileSize = 0;
+  long bytesRead = 0;
   long ftellFileSize = 0;
 
   /* check for path name */
